Added button group story
