+++ projects/ds-ng/src/lib/components/bmb-calendar/common/bmb-calendar-hour-view/bmb-calendar-hour-view.component.scss
@@ -2,7 +2,8 @@
 @use "../../../../../assets/styles/base/fonts" as fonts;
 
 .bmb_calendar-hours-row {
-  padding: var(--bmg-spacing-xs) var(--bmg-spacing-s) var(--bmg-spacing-xs) var(--bmg-spacing-m);
+  padding: var(--bmg-spacing-xs) var(--bmg-spacing-s) var(--bmg-spacing-xs)
+    var(--bmg-spacing-m);
   box-sizing: border-box;
   min-height: rem.rem_calc(60);
   border-bottom: rem.rem_calc(1) solid RGBA(var(--color-charade-200));
