+++ projects/ds-ng/src/lib/types/index.ts
@@ -1,5 +1,6 @@
 export type SizeNames = 'xs' | 's' | 'm' | 'l' | 'xl' | 'none' | 'auto';
-export type IButtonAppearance = 'primary'
-| 'secondary-filled'
-| 'secondary-outlined'
-| 'destructive';
+export type IButtonAppearance =
+  | 'primary'
+  | 'secondary-filled'
+  | 'secondary-outlined'
+  | 'destructive';
