+++ src/app/app.component.html
@@ -68,6 +68,8 @@
   </div>
 </div>
 
+<button bmbButton>Button 1</button>
+
 <div style="display: flex; gap: 3em">
   <a routerLink="/home" class="font-bold-6">Home</a>
   <a routerLink="/my-page" class="font-bold-6">My Page</a>
